Synced java.util.zip package with latest Android source. 	Change on 2014/03/10 by tball <tball@google.com> ------------- Created by MOE: http://code.google.com/p/moe-java MOE_MIGRATED_REVID=62865356

+++ jre_emul/android/libcore/luni/src/main/java/java/util/zip/Adler32.java
@@ -56,7 +56,7 @@ public class Adler32 implements Checksum {
      *            the byte to update checksum with.
      */
     public void update(int i) {
-        adler = updateByteImpl((byte) i, adler);
+        adler = updateByteImpl(i, adler);
     }
 
     /**
@@ -82,7 +82,7 @@ public class Adler32 implements Checksum {
         return adler32((uLong) adler1, (Bytef *) (buf->buffer_ + offset), (uInt) byteCount);
     ]-*/;
 
-    private native long updateByteImpl(byte val, long adler1) /*-[
+    private native long updateByteImpl(int val, long adler1) /*-[
         return adler32((uLong) adler1, (Bytef *) (&val), 1);
     ]-*/;
 }
